Add getAllUsers method to fetch user data from database + Add validateUser method for login credential verification + Add validateAdmin method for admin login checks
I have created a method called "getAllUsers" in the "dbReader" class. In order to interact with the database I used JDBC to establish a connection and execute queries. The logic I implemented allows for fetching and storing user data in an ArrayList. I also took care of handling SQLExceptions and providing feedback to the user.

To ensure password validation I introduced a method called "validateUser" in the "dbReader" class. This method establishes a database connection using JDBC. Prepares an SQL statement, for validation. It returns a value indicating whether the validation was successful or not while also including exception handling and feedback.

Additionally I added another method called "validateAdmin" in the "dbReader" class. This simple logic allows for validating hardcoded admin credentials. The method returns a value indicating whether or not the authentication of an admin is successful.